Take `playerCountScaling` field into account for holdout zones. Increase escape sequence timer.
Most holdout zones in the game are affected by the number of players within the radius. In fact,
the only outlier I have been able to identify is the pillars on the final stage. Regardless, we
should not be decreasing charge rate if the zone would not normally be affected by player count.

Additionally, it was possible - when configured to extreme values of "Additional Player Count"
and/or "Teleporter Duration" - for players to have insufficient time to charge the rescue ship
after defeating the final boss. As a result, the countdown timer will now scale proportionally
to these parameters.